add print_scene(), wip setting up dummy scene
in order to be able to develop the raytracing as
close as possible to the final version of input to
the functions i want to implement today and
tomorrow i set up some dummy objlst with a couple
of spheres in it. also a camera.

Also i made the print_scene func in order to check
if the data setup was successful.